VisionTrader/requirements.txt
pydantic>=2.6
mss>=9.0.1
Pillow>=10.2.0
numpy>=1.26
opencv-python>=4.9
pytesseract>=0.3.10
loguru>=0.7.2
requests>=2.31
pyautogui>=0.9.54
pyyaml>=6.0.1
